remove old tokenize
